Various improvements
Improved font rendering, improved save file handling, some settings sanitization, fixed a missing space in the settings menu background, can now back out of class selection menu if you entered it ingame

+++ code/InfiniminerClient/InterfaceItems/InterfaceButtonToggle.cs
@@ -81,7 +81,7 @@ namespace InterfaceItems
                 if (clicked)
                     dispText = onText;
 
-                spriteBatch.DrawString(uiFont, dispText, new Vector2(size.X + size.Width / 2 - uiFont.MeasureString(dispText).X / 2, size.Y + size.Height / 2 - 8), Color.Black);
+                spriteBatch.DrawString(uiFont, dispText, new Vector2(size.X + size.Width / 2 - (int)uiFont.MeasureString(dispText).X / 2, size.Y + size.Height / 2 - 8), Color.Black);
 
                 if (text != "")
                 {
